:bug: debug fake map

+++ path_searching/include/path_searching/path_node.h
@@ -37,7 +37,7 @@ template <typename T>
 struct MatrixHash : std::unary_function<T, size_t> {
   std::size_t operator()(T const& matrix) const {
     size_t seed = 0;
-    for (size_t i = 0; i < (long unsigned int)(matrix.size()); ++i) {
+    for (size_t i = 0; i < (matrix.size()); ++i) {
       auto elem = *(matrix.data() + i);
       seed ^= std::hash<typename T::Scalar>()(elem) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
     }
